Fix file location / installation
