+++ src/weiss_gripper_ieg76/serial_comm.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 import struct
 import time
 import serial
